+++ CHANGES.md
@@ -1,3 +1,9 @@
+1.0.0
+==========
+- A bunch of bug fixes
+- Change some prop names which may not be back-compatible
+- Pass all state of the table to any children component
+
 0.0.3
 ==========
 - update README